react: Split out <Results>
